fix: make "Old Kindle Mode" an opt-in feature via a button next to the home page logo
The button is hidden for modern browsers.

The last UA fix was a mistake, the UA of both the 2015 and 2021 Paperwhites are the same, despite their feature set being different. Maybe their JS-reported UA is still different, and that could be used instead with a reload + cookie compared to this temporary solution.